+++ hooks/use-scroll-to-bottom.ts
@@ -1,0 +1,44 @@
+import { useEffect, useRef, useState } from 'react'
+
+export const useScrollToBottom = <T>(dependency: T) => {
+  const listRef = useRef<HTMLDivElement>(null)
+  const scrollRef = useRef<HTMLDivElement>(null)
+  const [isAutoScroll, setIsAutoScroll] = useState(true)
+  const isScrolledByCode = useRef(false)
+
+  const handleScroll = (e: any) => {
+    if (isScrolledByCode.current) {
+      return
+    }
+    const { scrollTop, scrollHeight, clientHeight } = e.target
+    const isBottom = scrollHeight - scrollTop - clientHeight < 20
+    setIsAutoScroll(isBottom)
+  }
+
+  useEffect(() => {
+    if (listRef.current && isAutoScroll) {
+      isScrolledByCode.current = true
+      const timeoutId = setTimeout(() => {
+        if (listRef.current) {
+          listRef.current.scrollIntoView({
+            behavior: 'auto',
+            block: 'end',
+          })
+        }
+        // Reset the flag after scroll animation is likely finished
+        setTimeout(() => {
+          isScrolledByCode.current = false
+        }, 100)
+      }, 50)
+      return () => clearTimeout(timeoutId)
+    }
+  }, [dependency, isAutoScroll])
+
+  return {
+    scrollRef,
+    listRef,
+    handleScroll,
+    isAutoScroll,
+    setIsAutoScroll,
+  }
+}
